# backend/docs/swagger.yaml
basePath: /api
definitions:
  main.loginRequest:
    properties:
      identifier:
        type: string
      password:
        type: string
    type: object
  main.registerRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
  main.registeredUser:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API untuk aplikasi Barber One
  title: Barber One API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Kredensial
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - auth
  /logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Logout dan hapus sesi
      tags:
      - auth
  /profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Ambil profil user yang sedang login
      tags:
      - user
  /register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Data registrasi
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.registeredUser'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Registrasi user baru
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    description: 'Isi dengan: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"